+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Updated YouTube.js to v10
+- `Watch Later` and `Liked videos` are removed from "standard" items, as they are considered playlists.
+
 ## [0.4.7] - 2024-04-29
 
 ### Fixed